Hugging face/llamaindex_rag_agent/requirements.txt
datasets
langchain
langchain-community
huggingface-hub
smolagents
sentence-transformers
python-dotenv
#here we are using the environment variable is rag_agent

# Core libraries
numpy==1.25.2
pandas==2.1.1
datasets==2.13.2
huggingface_hub==0.17.1

# LangChain and community retrievers
langchain==0.1.0
langchain_community==0.0.159

# SmolAgents for agent creation
smolagents==0.1.11

# Optional for RAG embeddings / text processing
sentence-transformers==2.2.2
torch==2.3.1
transformers==4.33.2

# Visualization
IPython==8.16.2
Pillow==10.1.0
